BF-05: give the derived unit its BF16 contract and reference
The unit was copied from the FP32 one, so its ABI still said float32 everywhere.
Now q, k and v are BF16 in make_inputs and in validation, o is BF16 in the
contract outputs, in the stages block, in both references and in the pipeline's
output allocation, and the state and gates stay FP32.

The references widen the BF16 inputs once and run the same FP32 recurrence as
the FP32 unit, then round o, because that is what the kernel does. The domain is
unchanged from a5.gdn2_chunk_fwd, B=1 included, and the contract says so.

The comparison budget is the calibrated one: relative L2 at or below 5.2354e-03
= min(1e-2, 3F) for o and final_state, with the calibration cited in the reason
rather than restated.

reference passes all twelve cases; the simulator passes t1_h1_small end to end
including every stage checkpoint. Support stages stay untested until the runs
are recorded.

+++ kernels/projects/a5/gdn2_chunk_fwd_bf16/ref/stages.py
@@ -3,18 +3,18 @@ import torch
 
 
 def reference_stages(inputs):
-    q = inputs['q']
+    q = inputs['q'].float()
     batch, time, heads, _ = q.shape
     chunks = (time + 63) // 64
     qn = q * torch.rsqrt(q.square().sum(-1, keepdim=True) + 1e-6) * 128**-0.5
-    k = inputs['k']
+    k = inputs['k'].float()
     kn = k * torch.rsqrt(k.square().sum(-1, keepdim=True) + 1e-6)
     def pack(x):
         padding = torch.zeros(batch, chunks * 64 - time, heads, 128)
         return torch.cat((x, padding), dim=1).reshape(batch, chunks, 64, heads, 128).transpose(2, 3).contiguous()
     qn, kn, gc, bk, wv = (
         pack(qn), pack(kn), pack(inputs['g']).cumsum(-2),
-        pack(inputs['erase_gate'] * kn), pack(inputs['w'] * inputs['v']))
+        pack(inputs['erase_gate'] * kn), pack(inputs['w'] * inputs['v'].float()))
     lower = torch.zeros(batch, chunks, heads, 64, 64)
     score = torch.zeros_like(lower)
     for i in range(64):
@@ -34,7 +34,8 @@ def reference_stages(inputs):
         outputs.append((qn[:, c] * gc[:, c].exp()) @ state + score[:, c] @ delta)
         tail = kn[:, c] * (gc[:, c, :, -1:, :] - gc[:, c]).exp()
         state = gc[:, c, :, -1, :].exp().unsqueeze(-1) * state + tail.transpose(-1, -2) @ delta
-    o = torch.stack(outputs, dim=1).transpose(2, 3).reshape(batch, chunks*64, heads, 128)[:, :time].contiguous()
+    # o is the only BF16 checkpoint: the kernel writes it in BF16, so the reference rounds it too.
+    o = torch.stack(outputs, dim=1).transpose(2, 3).reshape(batch, chunks*64, heads, 128)[:, :time].contiguous().bfloat16()
     return dict(qn=qn, kn=kn, gc=gc, bk=bk, wv=wv, lower=lower, score=score,
                 u=u, wy=wy, states=torch.stack(states, dim=1),
                 delta=torch.stack(deltas, dim=1), final_state=state, o=o)
